fix code generation for vectorized systems
MonetDB for example sends whole vectors (numpy arrays) to the UDF.
Now there are two options: 1. the user knows this and handles the
vectors in the code, or 2. the user expects scalar types.

For 1, we don't need to do anything. For 2 we enclose the original
function in the generated UDF and call it for every element in the
vector.